+++ src/bit_iterator.rs
@@ -57,7 +57,6 @@ mod test {
     use super::*;
     use alloc::vec::Vec;
     use dusk_bls12_381::BlsScalar;
-    use dusk_bytes::Serializable;
 
     #[test]
     fn test_bit_iterator8() {