+++ src/util.rs
@@ -8,6 +8,7 @@ use alloc::vec::Vec;
 use dusk_bls12_381::{
     BlsScalar, G1Affine, G1Projective, G2Affine, G2Projective,
 };
+use ff::Field;
 use rand_core::{CryptoRng, RngCore};
 
 #[cfg(feature = "rkyv-impl")]
@@ -42,22 +43,17 @@ pub(crate) fn powers_of(
     powers
 }
 
-/// Generates a random BlsScalar using a RNG seed.
-pub(crate) fn random_scalar<R: RngCore + CryptoRng>(rng: &mut R) -> BlsScalar {
-    BlsScalar::random(rng)
-}
-
 /// Generates a random G1 Point using an RNG seed.
 pub(crate) fn random_g1_point<R: RngCore + CryptoRng>(
     rng: &mut R,
 ) -> G1Projective {
-    G1Affine::generator() * random_scalar(rng)
+    G1Affine::generator() * BlsScalar::random(rng)
 }
 /// Generates a random G2 point using an RNG seed.
 pub(crate) fn random_g2_point<R: RngCore + CryptoRng>(
     rng: &mut R,
 ) -> G2Projective {
-    G2Affine::generator() * random_scalar(rng)
+    G2Affine::generator() * BlsScalar::random(rng)
 }
 
 /// This function is only used to generate the SRS.